+++ tests/roots/test-base/target/configuration_settings_hide_reused_validator.py
@@ -1,0 +1,15 @@
+from pydantic import BaseModel, validator
+
+
+def validation(name):
+    """Validation function."""
+    return name
+
+
+class SettingOne(BaseModel):
+    name: str
+    """Name"""
+
+    normalize_name = validator('name', allow_reuse=True)(validation)
+    """Reused validator class method."""
+